+++ curseforge-client/src/models/fingerprint_match.rs
@@ -13,7 +13,7 @@ use crate::models;
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct FingerprintMatch {
     #[serde(rename = "id")]
-    pub id: u32,
+    pub id: u64,
     #[serde(rename = "file")]
     pub file: models::File,
     #[serde(rename = "latestFiles")]
@@ -21,7 +21,7 @@ pub struct FingerprintMatch {
 }
 
 impl FingerprintMatch {
-    pub fn new(id: u32, file: models::File, latest_files: Vec<models::File>) -> Self {
+    pub fn new(id: u64, file: models::File, latest_files: Vec<models::File>) -> Self {
         Self {
             id,
             file,
