+++ curseforge-client/src/models/minecraft_game_version.rs
@@ -13,9 +13,9 @@ use crate::models;
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct MinecraftGameVersion {
     #[serde(rename = "id")]
-    pub id: u32,
+    pub id: u64,
     #[serde(rename = "gameVersionId")]
-    pub game_version_id: u32,
+    pub game_version_id: u64,
     #[serde(rename = "versionString")]
     pub version_string: String,
     #[serde(rename = "jarDownloadUrl")]
@@ -27,7 +27,7 @@ pub struct MinecraftGameVersion {
     #[serde(rename = "dateModified")]
     pub date_modified: String,
     #[serde(rename = "gameVersionTypeId")]
-    pub game_version_type_id: u32,
+    pub game_version_type_id: u64,
     #[serde(rename = "gameVersionStatus")]
     pub game_version_status: models::GameVersionStatus,
     #[serde(rename = "gameVersionTypeStatus")]
@@ -35,7 +35,7 @@ pub struct MinecraftGameVersion {
 }
 
 impl MinecraftGameVersion {
-    pub fn new(id: u32, game_version_id: u32, version_string: String, jar_download_url: ::url::Url, json_download_url: ::url::Url, approved: bool, date_modified: String, game_version_type_id: u32, game_version_status: models::GameVersionStatus, game_version_type_status: models::GameVersionTypeStatus) -> Self {
+    pub fn new(id: u64, game_version_id: u64, version_string: String, jar_download_url: ::url::Url, json_download_url: ::url::Url, approved: bool, date_modified: String, game_version_type_id: u64, game_version_status: models::GameVersionStatus, game_version_type_status: models::GameVersionTypeStatus) -> Self {
         Self {
             id,
             game_version_id,
